fix: types in all files
