+++ src/common/types/coursesList.types.ts
@@ -1,4 +1,4 @@
-export interface Course {
+export interface ICourseList {
   id: string;
   title: string;
   tags: string[];
@@ -6,21 +6,47 @@ export interface Course {
   status: string;
   description: string;
   duration: number;
-  lessonsCount?: number;
+  lessonsCount: number;
   containsLockedLessons: boolean;
   previewImageLink: string;
   rating: number;
-  meta: Meta;
+  meta: IMeta;
 }
 
-export interface Meta {
+export interface ICourse {
+  id: string;
+  title: string;
+  tags: string[];
+  launchDate: string;
+  status: string;
+  description: string;
+  duration: number;
+  previewImageLink: string;
+  rating: number;
+  meta: IMeta;
+  lessons: ILesson[];
+  containsLockedLessons: boolean;
+}
+
+export interface IMeta {
   slug?: string;
   skills?: string[];
-  courseVideoPreview?: CourseVideoPreview;
+  courseVideoPreview?: ICourseVideoPreview;
 }
 
-export interface CourseVideoPreview {
+export interface ICourseVideoPreview {
   link: string;
   duration: number;
   previewImageLink: string;
+}
+export interface ILesson {
+  id: string;
+  title: string;
+  duration: number;
+  order: number;
+  type: string;
+  status: string;
+  link: string;
+  previewImageLink: string;
+  meta: IMeta;
 }
