--- src/common/types/types.ts
@@ -1,2 +1,0 @@
-export type { ICoursesList } from './coursesList.types';
-export type { ICourse } from './course.types';
