+++ src/components/courses/components/course-details/course-details.tsx
@@ -4,6 +4,13 @@ import { ENV } from 'common/enums/enums';
 import Spinner from 'components/common/loader/loader';
 
 import styles from './course-details.module.scss';
+import { ICourse } from 'common/types/coursesList.types';
+
+interface ICourseDetails {
+  response: ICourse | null;
+  error: Error | null;
+  loading: 'idle' | 'pending' | 'succeeded' | 'failed';
+}
 
 const CourseDetails: React.FC = () => {
   const { id } = useParams();
@@ -11,14 +18,15 @@ const CourseDetails: React.FC = () => {
     loading,
     response: course,
     error,
-  } = useFetch(`${ENV.API_PATH}/${id}`, 'id');
-  if (!course && loading === 'pending') return <Spinner isOverflowParent />;
-  if (!course && error)
+  } = useFetch(`${ENV.API_PATH}/${id}`, 'id') as ICourseDetails;
+
+  if (loading === 'pending') return <Spinner isOverflowParent />;
+  if (error)
     return <h3>{`Server response with  ${error?.toString()}`}</h3>;
 
   return (
     <div className={styles['course-page']}>
-      {course && <CourseInfo course={course} />}
+      {course ? <CourseInfo course={course} /> : null}
     </div>
   );
 };
