+++ src/components/courses/components/course-skills/course-skills.tsx
@@ -1,5 +1,5 @@
-import { IMeta } from 'common/types/course.types';
 
+import { IMeta } from 'common/types/coursesList.types';
 import styles from './course-skills.module.scss';
 
 interface IMetaProps {
